refactor(datasource/github-releases): Consolidate into single file
